+++ lib/handlers/pageHandler.js
@@ -4,6 +4,8 @@ const eventHandler = require('../eventBus');
 const logEvent = require('../logger');
 const nodeURL = require('url');
 const path = require('path');
+const { isSelector, elements } = require('../elementSearch');
+const fs = require('fs');
 let page, framePromises, frameNavigationPromise;
 
 eventHandler.on('createdSession', async (client) => {
@@ -74,6 +76,37 @@ const resolveFrameEvent = p => {
     }
 };
 
+const captureScreenshot = async (domHandler, selector, options) => {
+    if (selector && !isSelector(selector)) options = selector;
+    options.path = options.path || `Screenshot-${Date.now()}.png`;
+    let screenShot;
+    let clip;
+    if (isSelector(selector)) {
+        if (options.fullPage) console.warn('Ignoring fullPage screenshot as custom selector is found!');
+        let padding = options.padding || 0;
+        let elems = await elements(selector);
+        const { x, y, width, height } = await domHandler.boundBox(elems[0]);
+        clip = {
+            x: x - padding,
+            y: y - padding,
+            width: width + padding * 2,
+            height: height + padding * 2,
+            scale: 1
+        };
+        screenShot = await page.captureScreenshot({ clip });
+    } else if (options.fullPage) {
+        const metrics = await page.getLayoutMetrics();
+        const width = Math.ceil(metrics.contentSize.width);
+        const height = Math.ceil(metrics.contentSize.height);
+        clip = { x: 0, y: 0, width, height, scale: 1 };
+        screenShot = await page.captureScreenshot({ clip });
+    } else {
+        screenShot = await page.captureScreenshot();
+    }
+    if (options.encoding === 'base64') return screenShot.data;
+    fs.writeFileSync(options.path, Buffer.from(screenShot.data, 'base64'));
+};
+
 const handleNavigation = async (url) => {
     let resolveResponse, requestId;
 
@@ -109,4 +142,4 @@ const handleNavigation = async (url) => {
     eventHandler.removeListener('requestStarted', handleRequest);
 };
 
-module.exports = { handleNavigation, resetPromises };
+module.exports = { handleNavigation, resetPromises, captureScreenshot };
