+++ lib/taiko.js
@@ -20,7 +20,7 @@ const path = require('path');
 const eventHandler = require('./eventBus');
 const EventEmitter = require('events');
 const descEvent = new EventEmitter();
-let chromeProcess, temporaryUserDataDir, page, network, runtime, input, _client, dom, overlay, criTarget, 
+let chromeProcess, temporaryUserDataDir, page, network, runtime, input, _client, dom, overlay, criTarget,
     currentPort, currentHost, security, device, eventHandlerProxy, clientProxy, localProtocol = false;
 
 module.exports.emitter = descEvent;
@@ -355,7 +355,7 @@ module.exports.setViewPort = async (options) => {
  * @param {number} [options.navigationTimeout=5000] - Navigation timeout value in milliseconds for navigation after click. Accepts value in milliseconds.
  * @param {number} [options.waitForStart=100] - time to wait to check for occurrence of page load events. Accepts value in milliseconds.
  * @param {string[]} [options.waitForEvents = ['firstMeaningfulPaint']] - Page load events to implicitly wait for. Events available to wait for ['DOMContentLoaded', 'loadEventFired', 'networkAlmostIdle', 'networkIdle', 'firstPaint', 'firstContentfulPaint', 'firstMeaningfulPaint']]
- * 
+ *
  * @returns {Promise}
  */
 module.exports.openTab = async (targetUrl, options = { navigationTimeout: defaultConfig.navigationTimeout }) => {
@@ -566,7 +566,7 @@ module.exports.setLocation = async (options) => {
  * @param {number} [options.navigationTimeout=30000] - Navigation timeout value in milliseconds for navigation after click.
  * @param {Object} options.headers - Map with extra HTTP headers.
  * @param {number} [options.waitForStart = 100] - time to wait for navigation to start. Accepts value in milliseconds.
- * 
+ *
  * @returns {Promise}
  */
 module.exports.goto = async (url, options = { navigationTimeout: defaultConfig.navigationTimeout }) => {
@@ -594,7 +594,7 @@ module.exports.goto = async (url, options = { navigationTimeout: defaultConfig.n
  * @param {number} [options.timeout=30000] - DEPRECATED timeout option, use navigationTimeout instead.
  * @param {number} [options.navigationTimeout=30000] - Navigation timeout value in milliseconds for navigation after click.
  * @param {number} [options.waitForStart = 100] - time to wait for navigation to start. Accepts value in milliseconds.
- * 
+ *
  * @returns {Promise}
  */
 module.exports.reload = async (url, options = { navigationTimeout: defaultConfig.navigationTimeout }) => {
@@ -640,7 +640,7 @@ module.exports.goBack = async (options = { navigationTimeout: defaultConfig.navi
  * @param {number} [options.timeout=30000] - DEPRECATED timeout option, use navigationTimeout instead.
  * @param {number} [options.navigationTimeout=30000] - Navigation timeout value in milliseconds for navigation after click.
  * @param {number} [options.waitForStart = 100] - time to wait for navigation to start. Accepts value in milliseconds.
- * 
+ *
  * @returns {Promise}
  */
 module.exports.goForward = async (options = { navigationTimeout: defaultConfig.navigationTimeout }) => {
@@ -806,7 +806,7 @@ async function _click(selector, options, ...args) {
  * @param {string[]} [options.waitForEvents = ['firstMeaningfulPaint']] - Events available to wait for ['DOMContentLoaded', 'loadEventFired', 'networkAlmostIdle', 'networkIdle', 'firstPaint', 'firstContentfulPaint', 'firstMeaningfulPaint']
  * @param {number} [options.waitForStart=100] - time to wait for navigation to start. Accepts time in milliseconds.
  * @param {relativeSelector[]} args
- * 
+ *
  * @returns {Promise}
  */
 module.exports.click = click;
@@ -842,7 +842,7 @@ async function click(selector, options = {}, ...args) {
  * @param {Object} options
  * @param {boolean} [options.waitForNavigation=true] - Wait for navigation after the click. Default navigation timout is 30 seconds, to override pass `{ navigationTimeout: 10000 }` in `options` parameter.
  * @param {relativeSelector[]} args
- * 
+ *
  * @returns {Promise}
  */
 module.exports.doubleClick = async (selector, options = {}, ...args) => {
@@ -869,7 +869,7 @@ module.exports.doubleClick = async (selector, options = {}, ...args) => {
  * @param {selector|string} selector - A selector to search for element to right click. If there are multiple elements satisfying the selector, the first will be clicked.
  * @param {Object} options - Click options.
  * @param {boolean} [options.waitForNavigation=true] - Wait for navigation after the click. Default navigation timout is 30 seconds, to override pass `{ navigationTimeout: 10000 }` in `options` parameter.
- * 
+ *
  * @returns {Promise}
  */
 module.exports.rightClick = async (selector, options = {}, ...args) => {
@@ -897,7 +897,7 @@ module.exports.rightClick = async (selector, options = {}, ...args) => {
  * @param {selector|string} source - Element to be Dragged
  * @param {selector|string} destination - Element for dropping the dragged element
  * @param {Object} distance - Distance to be moved from position of source element
- * 
+ *
  * @returns {Promise}
  */
 module.exports.dragAndDrop = async (source, destination) => {
@@ -1018,7 +1018,7 @@ module.exports.focus = async (selector, options = {}) => {
  * @param {number} [options.timeout=30000] - DEPRECATED timeout option, use navigationTimeout instead.
  * @param {number} [options.navigationTimeout=30000] - Navigation timeout value in milliseconds for navigation after click.
  * @param {string[]} [options.waitForEvents = ['firstMeaningfulPaint']] - Events available to wait for ['DOMContentLoaded', 'loadEventFired', 'networkAlmostIdle', 'networkIdle', 'firstPaint', 'firstContentfulPaint', 'firstMeaningfulPaint']
- * 
+ *
  * @returns {Promise}
  */
 module.exports.write = async (text, into, options = { delay: 10 }) => {
@@ -1088,7 +1088,7 @@ const _write = async (text, activeElement, options) => {
  * @param {number} [options.timeout=30000] - DEPRECATED timeout option, use navigationTimeout instead.
  * @param {number} [options.navigationTimeout=30000] - Navigation timeout value in milliseconds for navigation after click.
  * @param {string[]} [options.waitForEvents = ['firstMeaningfulPaint']] - Events available to wait for ['DOMContentLoaded', 'loadEventFired', 'networkAlmostIdle', 'networkIdle', 'firstPaint', 'firstContentfulPaint', 'firstMeaningfulPaint']
- * 
+ *
  * @returns {Promise}
  */
 module.exports.clear = async (selector, options = {}) => {
@@ -1163,7 +1163,7 @@ module.exports.attach = async (filepath, to) => {
  * @param {number} [options.timeout=30000] - DEPRECATED timeout option, use navigationTimeout instead.
  * @param {number} [options.navigationTimeout=30000] - Navigation timeout value in milliseconds for navigation after click.
  * @param {string[]} [options.waitForEvents = ['firstMeaningfulPaint']] - Events available to wait for ['DOMContentLoaded', 'loadEventFired', 'networkAlmostIdle', 'networkIdle', 'firstPaint', 'firstContentfulPaint', 'firstMeaningfulPaint']
- * 
+ *
  * @returns {Promise}
  */
 module.exports.press = async (keys, options = {}) => {
@@ -1195,7 +1195,7 @@ async function _press(keys, options) {
  *
  * @param {selector|string} selector - A selector of an element to highlight. If there are multiple elements satisfying the selector, the first will be highlighted.
  * @param {relativeSelector[]} args - Proximity selectors
- * 
+ *
  * @returns {Promise}
  */
 module.exports.highlight = highlight;
@@ -1262,7 +1262,7 @@ async function mouseAction(action, coordinates, options = {}) {
  * @param {selector|string} selector - A selector of an element to scroll to.
  * @param {Object} options
  * @param {string[]} [options.waitForEvents = ['firstMeaningfulPaint']] - Events available to wait for ['DOMContentLoaded', 'loadEventFired', 'networkAlmostIdle', 'networkIdle', 'firstPaint', 'firstContentfulPaint', 'firstMeaningfulPaint']
- * 
+ *
  * @returns {Promise}
  */
 module.exports.scrollTo = async (selector, options = {}) => {
@@ -1316,7 +1316,7 @@ const scroll = async (e, px, scrollPage, scrollElement, direction) => {
  *
  * @param {selector|string|number} [e='Window']
  * @param {number} px - Accept px in pixels
- * 
+ *
  * @returns {Promise}
  */
 module.exports.scrollRight = async (e, px = 100) => {
@@ -1343,7 +1343,7 @@ module.exports.scrollRight = async (e, px = 100) => {
  *
  * @param {selector|string|number} [e='Window']
  * @param {number} px - Accept px in pixels
- * 
+ *
  * @returns {Promise}
  */
 module.exports.scrollLeft = async (e, px = 100) => {
@@ -1370,7 +1370,7 @@ module.exports.scrollLeft = async (e, px = 100) => {
  *
  * @param {selector|string|number} [e='Window']
  * @param {number} px - Accept px in pixels
- * 
+ *
  * @returns {Promise}
  */
 module.exports.scrollUp = async (e, px = 100) => {
@@ -1397,7 +1397,7 @@ module.exports.scrollUp = async (e, px = 100) => {
  *
  * @param {selector|string|number} [e='Window']
  * @param {number} px - Accept px in pixels
- * 
+ *
  * @returns {Promise}
  */
 module.exports.scrollDown = async (e, px = 100) => {
@@ -1422,42 +1422,17 @@ module.exports.scrollDown = async (e, px = 100) => {
  * await screenshot({fullPage:true})
  * await screenshot(text('Images', toRightOf('gmail')))
  *
- * @param {selector|string} selector 
+ * @param {selector|string} selector
  * @param {Object} options
- * 
+ *
  * @returns {Promise<Buffer>} - Promise which resolves to buffer with captured screenshot if {encoding:'base64'} given.
  */
 module.exports.screenshot = async (selector, options = {}) => {
     validate();
-    if (selector && !isSelector(selector)) options = selector;
-    options.path = options.path || `Screenshot-${Date.now()}.png`;
-    let screenShot;
-    let clip;
-    if (isSelector(selector)) {
-        if (options.fullPage) console.warn('Ignoring fullPage screenshot as custom selector is found!');
-        let padding = options.padding || 0;
-        let elems = await elements(selector);
-        const { x, y, width, height } = await domHandler.boundBox(elems[0]);
-        clip = {
-            x: x - padding,
-            y: y - padding,
-            width: width + padding * 2,
-            height: height + padding * 2,
-            scale: 1
-        };
-        screenShot = await page.captureScreenshot({ clip });
-    } else if (options.fullPage) {
-        const metrics = await page.getLayoutMetrics();
-        const width = Math.ceil(metrics.contentSize.width);
-        const height = Math.ceil(metrics.contentSize.height);
-        clip = { x: 0, y: 0, width, height, scale: 1 };
-        screenShot = await page.captureScreenshot({ clip });
-    } else {
-        screenShot = await page.captureScreenshot();
-    }
-    if (options.encoding === 'base64') return screenShot.data;
-    fs.writeFileSync(options.path, Buffer.from(screenShot.data, 'base64'));
-    descEvent.emit('success', 'Screenshot is created at ' + options.path);
+    console.log('----', options);
+    pageHandler.captureScreenshot(domHandler, selector, options);
+    console.log(options);
+    descEvent.emit('success', 'Screenshot is created at ' + options.encoding ? options.encoding : options.path);
 };
 
 /**
@@ -1468,10 +1443,10 @@ module.exports.screenshot = async (selector, options = {}) => {
  *
  * @param {selector} selector
  * @param {Object} options
- * @param {boolean} [options.waitForNavigation = true] - - Wait for navigation after the click. Default navigation timeout is 15 seconds, to override pass `{ navigationTimeout: 10000 }` in `options` parameter. 
+ * @param {boolean} [options.waitForNavigation = true] - - Wait for navigation after the click. Default navigation timeout is 15 seconds, to override pass `{ navigationTimeout: 10000 }` in `options` parameter.
  * @param {string[]} [options.waitForEvents = ['firstMeaningfulPaint']] - Events available to wait for ['DOMContentLoaded', 'loadEventFired', 'networkAlmostIdle', 'networkIdle', 'firstPaint', 'firstContentfulPaint', 'firstMeaningfulPaint']
  * @param {relativeSelector[]} args - Proximity selectors
- * 
+ *
  * @returns {Promise}
  */
 module.exports.tap = async (selector, options = {}, ...args) => {
@@ -1595,7 +1570,7 @@ module.exports.listItem = (attrValuePairs, ...args) => {
 /**
  * This {@link selector} lets you identify a button on a web page with label or attribute and value pairs and proximity selectors.
  * Tags button and input with type submit, reset and button are identified using this selector
- * 
+ *
  * @example
  * await highlight(button('Get Started'))
  * await button('Get Started').exists()
